Import finalizer logic from miniscript
Import the `psbt` module from `rust-miniscript` while tip of master was:

  a280447 Merge rust-bitcoin/rust-miniscript#617: Improve `TapTree` API

Only do changes required to get things building, however this includes
upgrade of the bitcoin dependency.

Feature gate the miniscript stuff because we want this crate to be a
drop in replacement for the `psbt` module in `rust-bitcoin`, users may
not want to have `rust-miniscript` in there dependency tree.