Add a PSBT v0 2-of-2-multisig example
Add an example of creating a 2-of-2 multisig output using the `v0::Psbt`
API. Currently just creates the mutisig output from two p2wpkh UTXOs,
does not spend the output.

Note this example does not use miniscript to finalize and I do not even
know if that is a sane thing to expect.