v0: Fail to combine non-v0 PSBTs
It does not make sense to allow our `v0` implementation to combine two
PSBTs with different versions, error if this is attempted.

While we are at it, move the `CombineError` into `global` because that
is where it originates. Although this is not important in `v0`, over in
`v2` there will be different `CombineError`s for the different maps -
lets be uniform already.